fixed issue with event log file being written to program root
