+++ faultinject-bundle/src/test/java/org/intocps/maestro/faultinject/InjectCorrectnessTest.java
@@ -24,7 +24,7 @@ public class InjectCorrectnessTest {
         final File faultInjectSpec = Paths.get("target", "funcevaltestcorrect", "FaultInject.mabl").toFile();
         faultInjectSpec.getParentFile().mkdirs();
         try (final FileWriter writer = new FileWriter(faultInjectSpec)) {
-            IOUtils.copy(FaultInjectRuntimeModule.class.getResourceAsStream("FaultInject.mabl"), writer, StandardCharsets.UTF_8);
+            IOUtils.copy(FaultInjectLivecycleHandler.class.getResourceAsStream("FaultInject.mabl"), writer, StandardCharsets.UTF_8);
         }
 
         final File spec = Paths.get("target", "funcevaltestcorrect", "funceval_output_correctness_test.mabl").toFile();
